+++ src/main/kotlin/me/vendoor/dragonball/api/migration/MigrationContext.kt
@@ -1,7 +1,7 @@
 package me.vendoor.dragonball.api.migration
 
 import com.mongodb.client.MongoDatabase
-import me.vendoor.dragonball.api.dsl.creation.CreateCollectionSpecificationContextBuilder
+import me.vendoor.dragonball.api.dsl.upsert.CreateCollectionSpecificationContextBuilder
 
 class MigrationContext(private val database: MongoDatabase) {
     fun createCollections(lambda: CreateCollectionSpecificationContextBuilder.() -> Unit) {
