+++ src/main/kotlin/me/vendoor/dragonball/api/setup/SetupCreationDslProcessor.kt
@@ -6,9 +6,9 @@ import com.mongodb.client.MongoDatabase
 import com.mongodb.client.model.CreateCollectionOptions
 import com.mongodb.client.model.IndexOptions
 import me.vendoor.dragonball.api.configuration.Configuration
-import me.vendoor.dragonball.api.dsl.creation.CreateCollectionContext
-import me.vendoor.dragonball.api.dsl.creation.CreateDatabaseContext
-import me.vendoor.dragonball.api.dsl.creation.CreateIndexContext
+import me.vendoor.dragonball.api.dsl.upsert.CreateCollectionContext
+import me.vendoor.dragonball.api.dsl.upsert.CreateDatabaseContext
+import me.vendoor.dragonball.api.dsl.upsert.CreateIndexContext
 import me.vendoor.dragonball.api.migration.MigrationPerformer
 import me.vendoor.dragonball.api.util.database.hasCollection
 import org.bson.BsonDocument
