+++ src/main/kotlin/me/vendoor/dragonball/specification/VendoorDatabaseSpecification.kt
@@ -1,6 +1,6 @@
 package me.vendoor.dragonball.specification
 
-import me.vendoor.dragonball.api.dsl.creation.database
+import me.vendoor.dragonball.api.dsl.upsert.database
 
 fun vendoorDatabaseSpecification() = database {
     name { "vendoor" }
